Remove lib/security.lib.php
Unless we're going to commit to providing a stable, machine-friendly
way of retrieving the latest version, we shouldn't pretend that we can
check it in an automated fashion.

Nor should we attempt to do so for PHP.  Users are responsible for their
security; we only care if their PHP installation provides the features
we need in order to work.