Clean up xmlData::songs
Drop call to Song->format(); two of the values that we use are exposed
via public getters anyway, and it's not particularly ugly to call
Tag::get_top_tags instead of relying on Song to do that for us.

Remove genre cruft.

Based on a patch by lotan_rm in FS#215