Refactor similar init code into new init-tiny.php
Several places were doing operations that were practically the same
(loading core libraries and setting some variables), with varying
levels of success in remembering to do the right things in the right
order. Making a separate minimal init.php that doesn't worry about
some of the sanity checks and session bumpf makes this cleaner and
easier to maintain.