Fix pluralisation issue in Recently Played
Previously we would say things like "1 seconds" and "1 hours". Using
the magic of ngettext and sprintf we can get rid of the static
concatenation order and become theoretically more translatable as well
as more correct in English.

Also changed the logic a bit, since the old looping method was
unfriendly to my brain.