FS#142 - Clicking on song results in downloading
Change the title link in a song row so it links to the song details.

While there's no real value in obfuscating the fact that a stream
can be downloaded, there is value in a consistent interface. The
song title being a stream link is inconsistent with other parts of
the row (such as album and artist), inconsistent with other times we
display songs (such as Recently Played), and doesn't consistently match
the currently selected playback method.