Clean up Content-Range handling for playback
Transcoding can't cleanly support Content-Range, so ignore it if we
receive it and send 'Accept-Ranges: none' so smart clients won't send
it.

The logic that determines when enough has been streamed that the
stream should be recorded in the stats was complexified. Since we've
given up on predetermining the file size for transcodes we have to
account for that. Content-Ranges that start well into the file now
won't update the stats, since we've probably already done it (this
assumption is wrong if the client immediately seeks, but nothing's
perfect.)  Finally, a straight file size / 4 seemed too straighforward,
so small files will be / 1.1 and large files will have the value capped.