Mess around with ratings
Drop the public variables from the rating class; everyone should use the
getters.

Add the ability for themes and applications to distinguish between
a user's actual rating and the global average rating; in the web
interface the average shows up if a user hasn't rated something but at
least one other user has.