Change css handling for RTL languages
Code duplication is bad even when it's CSS, mmkay?

If a theme has templates/rtl.css, it will be loaded in addition to
default.css.  This is more sane than the old behaviour, which loaded
default-rtl.css instead of default.css.

I'm not feeling ambitious enough to actually check whether the current
-rtl.css files work and change them to overrides instead of modified
copypasta, so I just dropped them for now.