install: Fix GRANT for non-localhost servers
If we're not connecting locally, tell MySQL to allow our new user to
connect from any host. Previously we restricted them to connecting
from the database host, which is clearly wrong.

It would be slightly better to properly restrict it to the Ampache
host, but let's not overcomplicate the install and confuse the users.