Added cookies to remember how the collapsed menu were placed.
It automatically replaces the menu based on the cookie value. Also
the icon is placed accordingly. The JS is in the sidebar.inc.php
file due to the fact that changing from home to admin for instance
is actually an Ajax call and if the JS is not within the content
rendered with the Ajax call, it no longer works.

I couldn't get it to work for the Filters menu, as it's no rendered
by the sidebar.