Changed URL of the browse menu for Ajax calls.
Added the content.ajax.php that handles the actions from the sidebar
URLs. It calls the browse.php file as this browse.php handles all
that is needed to render the content.

Disabled the top and bottom boxes of the UI when it's an Ajax request.

For some reason, calling the Tags Cloud in Ajax like that breaks the
cloud display, hence why I didn't change this URL.