Modified all the sidebar URLs to AJAX calls.
All the sidebar URLs have been changed to AJAX calls but one. The tag
Cloud is still a normal URL, as I didn't find a fix for the display
issue that occurs when changing it.

A new static function has been made to quickly create Ajax links with
the page, subpage, action, type, etc.

The content.ajax.php has been changed so a much simpler way. Since
all the pages are already doing the rendering work, there is a switch
on the subpage that simply calls the right page and that's it.