Transformed the Ajax::text(Ajax::make_url()) into UI::create_link
The UI::create_link uses the Ajax::text(Ajax::make_url()) methods to
create all that is needed.

The Ajax::make_url() has been changed to take an array of key value
as a third parameter to allow a more flexible URL generation. However,
it requires to send a key => valuea array to the function call.