Fixed the Album of the Moment not displaying.
The Album of the Moment is actually a Jquery function that usually
triggers at the end of a page load. It wasn't triggered at the end
of an Ajax load. To have it triggered, the event ajaxComplete had
to be used.

To make sure the ajaxComplete event wasn't triggered for every
ajax call, the event listener has to be unbind after the first
call, using the $().off(ajaxComplete) function.