Moved the load time to level 5 debug event.
Moved the load time, queries and cache hits from the footer to the
end of content.ajax.php. Made it a level 5 debug event.

Changed that since the Ajax call won't refresh the value from the
footer. I believe the footer could be changed to an event listener
that refreshes after each ajax call, but considering the amount of
ajax calls, the result might not be accurate/worth metionning.