Ajax calls added to header, and new sidebar links.
Links in the Header have been changed for Ajax calls. The quick
search function has been changed. The auto-complete function has
been modified to take ajax call URLs. It currently only works
for the song but the changes to the other objects will work for
this later on.

The new addition to the sidebar has been changed for Ajax calls. It's
the licenses and the upload that have been added.

The song class has been altered so the four types of link it contains
are ajax calls.