Ger Translation + Translation Catalog updated. + catalog_update.inc
I sometimes ran into errors if I import all my catalogs, because of the
PHP memory limit. So I decided to add an option for a temporary
deactivated PHP-memory limit. it is default set to off, so there is
enough security for user who read the help^^
I also have added the option to decide if you want a Database Table
optimization or not, because in my case it takes about 5 min to complete
and I wont ever optimize my tables.

last but not least:
I have disconnected the forced gather-art and import playlist from local
catalog add. Sometimes you wanna add songs as fast as possible. Gather
Art has no priority für a funktioning and useful Media library, so it
could be done after it or if the user wants it, it can also be done via
a switch. Like all the others.

So I think I've mentioned all changes I've made^^

EDIT: Updated (7)
EDIT: There isn't a -t (any more)