A small change to make translation easier
It's always better to keep those variable user-, plugin-, etc-names as
dynamic as possible. Otherwise it's sometimes extremely problematic to
find the right sentence-formatting. In German I would translate it that
way:

"Klicke um `$plugin_name` den Zugriff zu erlauben"

And does this English sentence in the source fit better? As I see from
the context, the user is asked to give `$plugin_name`the right to access
Ampache, is this right?

:-)