New position for update notification:
Since ever, I hated it that the Update notification is within the
headerbox. Now it's on the top of the content div like the "updated
config" notification.

I also kicked out one permission check so the
"autoupdate" notification and/or the "new config" notification is only
shown if the permission is 100... Like before but in another way.
^^

Before, there was a `<span>...</span>`, no matter what user was
logged in. Now there is a div with `id="update_notify"` what is also
only shown, if the admin is logged in.

I also cleaned the depending CSS
rule in `/reborn/templates/default.css`