Fix proposition for issue #123.
This commit introduces a `nT_` function to handle pluralization in
localizations, and avoid the need to call ngettext, which was resulting
in errors if PHP Gettext module was not available (and this was not
enforced by Ampache).

Any occurrence of `ngettext` has been replaced by this `nT_` call.

Also removed the useless keywords in the `gather-messages.sh` script,
and added the newly introduced `nT_` keyword.

Closes issue #123.