Fix for issue #1248 Generate `sceret_key` on Ampache installation
Generate a `secret_key` on Ampache installation and put it in the
generated config. This removes the need to do it manually, and the
secret key is ensured to be crypto safe.

Also move the `web_path` configuration setting under a "Various"
fieldset, as it had nothing to do with database settings.

Also added a comment about a missing `break` statement, which was on
purpose, but one could think it was a mistake at first sight.

Closes #1248.