Add opus transcoding arguments to ampache.cfg
Added a line for ffmpeg opus encoding (transcoding) using libopus (which obviously needs to be installed) to transcode on the fly to the very fast and efficient opus codec. (Very personal) bitrate recommendation: 128k (To be set in the the admin interface or user personal preferences).
I'm suggesting ogg as a transport (ogg opus) which in turns requires the vsync 2 option.