Small changes to make translation easier
Changed and combined some strings to make translation in Transifex
easier. It's always better to not split up translatable strings, so
translators can see how they're put together and don't have to search
for the parts.

Also, in the `dropbox.catalog.php`, I changed the strings to not behave
like html tags. Transifex doesn't like this / treats it as html tags.
;-)

:-)