Latest translation updates and maybe fix for #1738
Finally I had some time and motivation to look into #1738 again. And I guess I fixed it. But the problem that the subcategory - database entries are processed somewhere to be written with first letter uppercase, (I still can't find it) could make translations ineffective.
I also see some conflicts to already existing translations for the word "upload".

The bad news is that we peobably have to rewrite the translation process to be more modular and get rid of these descriptions in the database and put them into the modules, or where ever they belong to, directly.

But I guess this shouldn't be done for this version of Ampache anymore. But maybe we (by that I mean @wagnered, because I don't even have an idea where to start :P) fix that with the new Laravel version.
Do you agree @wagnered?

:)