Bug:Change AlbumArtist + custom metadata
I changed the order in which the Metadata is parsed in write_id3. This will set custom fields to values of song-instance as they are defined in get_metadata, overriding them as needed. Otherwise the standard metadata like "band" will be overridden by custom value in each song that cannot be changed from edit album. So changing Albumartist wouldn't work.