Fix the layout when displaying images and some are filtered out.
The logic to display each image in the table cells was broken in the
case some of the images where filtered out due to size constraints.
The new logic simply keeps a counter for the table cell that
gets increased only when an image is added to a cell.
After 4 cells are filled a new row is created.