Recently played RSS feeds fixed for users who don't share played info
The recently played RSS feeds contain now a unique random
RSS token that allows to identify which user it refers to.
That way the feed, which has been obtained within the authenticated
page, contains the user recently played items, which wasn't the
case before.

The RSS token can be reset by the administrator, which will
modify the behaviour of previous feed URLS.