Make http_host auto-estimation more robust
Prefer HTTP_HOST over SERVER_NAME, if defined. SERVER_NAME is not reliable as it depends on the webserver config and might be empty, set to the local server hostname, a wildcard/regex or another invalid string for internal identification only. HTTP_HOST matches the client request and should hence work the same way the current client access worked, when using for redirects. If in unexpected circumstances both variables are not set or empty, "localhost" is used as fallback.

Signed-off-by: MichaIng <micha@dietpi.com>