Drop bogus session based error-reinsert in AmpError
Well, turns out: Did not work in develop and was `accidentally fixed` in
source-changes. The old `auto_init` autoloading did not work as expected
in here, so errors from previous requests were withdrawn. Explicitly
calling init now adds the errors to the internal error list, so they show up
multiple times if the same error occurs again. So, just drop the
initialization to preserve the behaviour of develop.