Don't try to connect to db if hostname is missing #2808
Due to some calls deep within ampache, the system attempts a db
connection although the config is not loaded at this point.
This causes some error entries in the mysql server log if the server is
localhost as the dsn fallbacks to `mysql:host=` (which seems to fallback
to localhost`).
Additionally:
- The conditional check for the hostname `$hostname ?: 'localhost'` does not work
because the string concatenation has higher precedence than the condition.
- No errors are logged as the logger is not initialized yet (also due to
  missing configuration)