Drop memory cache implementation
For most users, the memory cache wasn't a thing at all. Some may have
used it, but due to the nature of the php request handling, the cache
was not preserved between request. The new redis implementation was
introduced in edge and therefor not in use.

Caching will be re-implemented by the orm once we finish the db
refactoring. The orm will implicitly cache all entities within a request
and may explicitly cache query-results if enabled.