Catalog: Stop filtering items beginning with a "."
This is preventing legitimate directories and files from being scanned
into the catalog. ("Legitimate" here meaning files and directories where
the name matches that of the album or track, such as "...Like
Clockwork".

This style of skipping the pseudo-directories is consistent with that of
src/Repository/Model/Art.php.

fixes #3422