Provice a non-static database access
Static methods are hard to test/not testable, therfor we need a solution for quering the database
using non-static methods.

The DatabaseConnection-class will provide non-static methods - currently only `query` is available.
More methods will follow on a on-demand basis.

Use the new DatabaseConnection within the ShoutRepository and add tests for all existing methods.