Correct some incomplete type definitions #3671
Most properties in all of the model files are nullable but their doctype definition doesn't respect that.
So add a mitigation for certain locations for now to fix a recent issue.

In the long term, the docblock typehints of the properties should be replaced by native ones and they should also be made nullable.