Prevent the playback from temporarily updating media properties
Properties like `file`, `size` or `f_file` got temporarily overwritten by some catalog types during playback preparations.
This should not happen as those properties represent the internal state of the object and are not meant to be some sort of
`data storage` to carry some temporary information from place a to b.

I tested this with certain configuration but this surely somewhat experimental...