Allow the user api method to return the current user info
If a user uses the Api via Api-Key, the user name of the user is not known. Api methods that require the specification of a user name cannot be used in this way.

This change allows the `username` parameter to be omitted and in this case returns all the information of the currently used API user - including the username.