Implement a generic loader for library items
This is the successor to ObjectToClassNameMapper and enables generic library objects to be loaded using object-type and object-id tuples. Restrictions can be made using class/interface names.

On this occasion, an enum for all types of library items is also introduced and the whole thing is first used in the share object.